pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "smart_inbox_lite"
version = "1.0.0"
description = "A professional 3-task email triage environment with normalized scoring."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic",
    "openai",
    "python-dotenv",
    "requests",
    "httpx",
    "fastapi",
    "uvicorn",
    "openenv-core>=0.2.0"
]

[project.scripts]
server = "server.app:main"
inference = "inference:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
